Move plotting-specific data structures into corresponding module
